requirements.txt
couchdb
Mastodon.py
requests